fix(husky): do not stage all files from index
The current setup makes it such that whenever a commit is made, husky will go ahead and add all changes to staging as a part of the pre-commit hook, regardless of whether the user selected a particular set of files. This effectively removes the ability to cherry-pick files and push those only. 

This PR addresses that issue, especially when users might not want to push files associated with an unfinished change.