Add customizable Layout component for auth pages
Created a `Layout` component to provide customizable layouts for
login and signup pages. The available layouts in
`authentication-form-layout.tsx` are:

- `full-form`
- `half-image`
- `background-image`
- `default` (fallback)

The image for `background-image` and `half-image` layouts can be
customized in the assets folder.

Also made form props toggleable in login and signup forms, allowing
visibility control for email and password fields using
`showEmail=true` and `showPassword=true`.